Fixed the null-selector in RefactorToSatisfyDependencies
The bug was happening because we used Declaration#DeclarationEquals with
which may not always be able to find the declaration we are looking for.
For instance, when refactoring, we choose declaration with the smallest
number of chars to be included in the created grouping selector. In this
case, if there is a dependency to or from an equivalent declaration for
which Declaration#declarationEquals returns false, we cannot find the
declaration we are looking for.