Fixed a bug in virtual shorthand declaration creation
We now check if the new virtual shorthand styles the same properties
compared to the real individual declarations. We avoid grouping, for
instance, margin-left and margin-top in a virtual shorthand, while it
was possible in the past (margin-right and margin-bottom should be
present in the same selector).

The implementation is a hack, BTW (refer to the comments in the code)