Several fixes for adding necessary declarations and setMixinName
Fixed getDeclarationsToBeAdded(), However, there should be a method to
position the added declaration in the right place. 

There is no originalSelector anymore, so we fixed it in
LessMixinMigrationOppotunity.

In addition, setMixinName() now throws IllegalArgumentException in case
of an invalid mixin name. This might change in the future, to check for
duplicate names